fix: ensure onQueryCacheReturn all works when query needs call
This fixes an issue where a query would return only a subset of the
cache, but "all" was specified on the cache return. Now, when a
query is needed to be made, the response is used to update the
cache and then return all object from the cache (including those
not updated).